# apispec.json
{
  "components": {
    "schemas": {
...
      "ListMeta": {
        "properties": {
          "count": {
            "example": 20,
            "type": "integer"
          },
          "limit": {
            "example": 20,
            "type": "integer"
          },
          "page": {
            "example": 1,
            "type": "integer"
          },
          "total": {
            "example": 125,
            "type": "integer"
          },
          "total_pages": {
            "example": 7,
            "type": "integer"
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "total_pages",
          "count"
        ],
        "type": "object"
      },
...
      "CreateHadiRequest": {
        "properties": {
          "description": {
            "example": "A renowned Quran reciter",
            "nullable": true,
            "type": "string"
          },
          "image_url": {
            "example": "https://example.com/hadi.jpg",
            "format": "uri",
            "nullable": true,
            "type": "string"
          },
          "name": {
            "example": "Mishary Rashid Alafasy",
            "maxLength": 255,
            "minLength": 2,
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
      "UpdateHadiRequest": {
        "properties": {
          "description": {
            "example": "A renowned Quran reciter",
            "nullable": true,
            "type": "string"
          },
          "image_url": {
            "example": "https://example.com/hadi.jpg",
            "format": "uri",
            "nullable": true,
            "type": "string"
          },
          "name": {
            "example": "Mishary Rashid Alafasy",
            "maxLength": 255,
            "minLength": 2,
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
      "HadiResponse": {
        "properties": {
          "created_at": {
            "example": "2025-12-01T10:00:00Z",
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "example": "A renowned Quran reciter",
            "nullable": true,
            "type": "string"
          },
          "id": {
            "example": 1,
            "type": "integer"
          },
          "image_url": {
            "example": "https://example.com/hadi.jpg",
            "format": "uri",
            "nullable": true,
            "type": "string"
          },
          "name": {
            "example": "Mishary Rashid Alafasy",
            "type": "string"
          },
          "updated_at": {
            "example": "2025-12-01T10:00:00Z",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
      "HadiListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/HadiResponse"
            },
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ListMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "JWT token authentication. Use access_token from login response.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API specification for the Ishari backend service.",
    "title": "Ishari Backend API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/hadis": {
      "get": {
        "operationId": "listHadis",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HadiListResponse"
                }
              }
            },
            "description": "List of hadis"
          }
        },
        "summary": "List hadis",
        "tags": [
          "Hadis"
        ]
      },
      "post": {
        "operationId": "createHadi",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHadiRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/HadiResponse"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Hadi created successfully"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create hadi",
        "tags": [
          "Hadis"
        ]
      }
    },
    "/hadis/{id}": {
      "delete": {
        "operationId": "deleteHadi",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hadi deleted successfully"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Delete hadi",
        "tags": [
          "Hadis"
        ]
      },
      "get": {
        "operationId": "getHadiById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/HadiResponse"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Hadi details"
          }
        },
        "summary": "Get hadi by ID",
        "tags": [
          "Hadis"
        ]
      },
      "put": {
        "operationId": "updateHadi",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHadiRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/HadiResponse"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Hadi updated successfully"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update hadi",
        "tags": [
          "Hadis"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local development server",
      "url": "http://localhost:3000/api"
    }
  ]
}